Apply Val's patch to fix test errors

+++ unit_tests/examples/ex_cl_cvdunpack_test.py
@@ -8,6 +8,8 @@ import os
 import platform
 import shutil
 import sys
+from pathlib import Path
+
 
 sys.path.append('../unit_tests')
 import testcase
@@ -24,14 +26,9 @@ class TC(testcase.TestCase):
         super(TC, cls).setUpClass()
 
         # Find the example program
-        program_path = os.environ["EX_CL_CVDUNPACK_TEST"]
-
-        shutil.copy(
-            str( program_path ),
-            str(TC.path_build / 'unit_tests' / ( program_name + '.exe' ) ),
-        )
-
-        TC.example_program = TC.path_build / 'unit_tests' / ( program_name + '.exe' )
+        TC.example_program = Path(os.getenv("EX_CL_CVDUNPACK"))
+        if not TC.example_program.exists():
+            raise Exception(f'Could not find the example program {TC.example_program}')
 
         # Copy the test cvd to the temp directory
         shutil.copyfile(
